# src/main/openapi/university-openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
paths:
  '/api/courses/{id}':
    put:
      tags:
        - course-controller
      operationId: modifyCourse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
  /api/courses:
    get:
      tags:
        - course-controller
      operationId: getAllCourses
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
    post:
      tags:
        - course-controller
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
  '/api/courses/{id}/versions':
    get:
      tags:
        - course-controller
      operationId: getVersionAt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: at
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
  '/api/courses/{id}/history':
    get:
      tags:
        - course-controller
      operationId: getCourseHistoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryDataCourseDto'
  /api/courses/search:
    get:
      tags:
        - course-controller
      operationId: search
      parameters:
        - name: full
          in: query
          required: false
          schema:
            type: boolean
        - schema:
            type: integer
          in: query
          name: id
        - schema:
            type: string
          in: query
          name: name
        - schema:
            type: string
          in: query
          name: teachers.name
        - schema:
            type: string
          in: query
          name: students.id
        - schema:
            type: array
            items:
              type: integer
          in: query
          name: students.semester
        - schema:
            type: integer
          in: query
          name: page
        - schema:
            type: integer
          in: query
          name: size
        - schema:
            type: array
            items:
              type: string
          in: query
          name: sort
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
components:
  schemas:
    CourseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentDto'
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/TeacherDto'
    StudentDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        semester:
          type: integer
          format: int32
    TeacherDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
    HistoryDataCourseDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CourseDto'
        revisionType:
          type: string
          enum:
            - ADD
            - MOD
            - DEL
        revision:
          type: integer
          format: int32
        date:
          type: string
          format: date-time